fixed 500 errors on /admin/projects POST and PUT.
The error "The type 'List' used in the 'for' loop must implement
Iterable with a type argument that can be assigned to 'int'" started
to appear with a recent version of Dart and was causing 500 errors.
Adding a type to the imageIds List variable did not fix the problem, so
I kept the List type dynamic and casted the usage of each element in
the list to an int. TODO: Why doesn't setting the List type work?