Fixed uncolored theme switch on mobile
The color of the theme toggle switch intermittently does not get set on
initial page load. I believe this is happening when the CSS hasn't fully
loaded in time to be accessed through JS. To completely avoid this, the
colors have just been hardcoded into the config of the Switch component.